Fix issue with LRTDP. Exposed by changes to Multinomial.
The bug was entries in the policy weren't being filled. It turns out
this happens because we only fill in the policy at visited states.
For some graphs it's possible to solve a state without visiting it
when all subsequent states have become solved (if there are other
routes to reach that state).