Clean up multi-agent support
The current implementations of multi-agent learning are not
fully tested or entirely consistent with the rest of the
package. It also relies on the `sparse` library, which
has conflicts with the most updated version of `numpy`
via `numba`. For simplicity, I am removing references to
multi-agent parts of the library that cause problems
and including a warning when `TabularStochasticGame`
is called.